+++ tests/app/server.js
@@ -57,10 +57,10 @@ app.all('/links/preserve-state-page-two', (req, res) =>
   inertia.render(req, res, { component: 'Links/forgetStat', props: { foo: req.query.foo } }),
 )
 app.all('/links/preserve-scroll-page-two', (req, res) =>
-  inertia.render(req, res, { component: 'Links/forgetScroll', props: { foo: req.query.foo } }),
+  inertia.render(req, res, { component: 'Links/preserveScroll', props: { foo: req.query.foo } }),
 )
 app.all('/links/preserve-scroll-false-page-two', (req, res) =>
-  inertia.render(req, res, { component: 'Links/forgetScrollFalse', props: { foo: req.query.foo } }),
+  inertia.render(req, res, { component: 'Links/preserveScrollFalse', props: { foo: req.query.foo } }),
 )
 app.get('/links/as-warning/:method', (req, res) =>
   inertia.render(req, res, { component: 'Links/AsWarning', props: { method: req.params.method } }),
@@ -88,10 +88,10 @@ app.all('/visits/preserve-state-page-two', (req, res) =>
   inertia.render(req, res, { component: 'Visits/forgetState', props: { foo: req.query.foo } }),
 )
 app.all('/visits/preserve-scroll-page-two', (req, res) =>
-  inertia.render(req, res, { component: 'Visits/forgetScroll', props: { foo: req.query.foo } }),
+  inertia.render(req, res, { component: 'Visits/preserveScroll', props: { foo: req.query.foo } }),
 )
 app.all('/visits/preserve-scroll-false-page-two', (req, res) =>
-  inertia.render(req, res, { component: 'Visits/forgetScrollFalse', props: { foo: req.query.foo } }),
+  inertia.render(req, res, { component: 'Visits/preserveScrollFalse', props: { foo: req.query.foo } }),
 )
 app.post('/visits/events-errors', (req, res) =>
   inertia.render(req, res, { component: 'Visits/Events', props: { errors: { foo: 'bar' } } }),